requirements.txt
Flask
Flask-Cors
PyJWT
Werkzeug    